% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bess_svm.R
\name{bess_svm}
\alias{bess_svm}
\title{Best subset selection on svm.}
\usage{
bess_svm(X, y, T0, alpha = 0.01, tau = 0.05, max.steps = 100)
}
\arguments{
\item{X}{Data Matrix}

\item{y}{True label}

\item{T0}{Fixed nonzero feature number}

\item{alpha}{Penalty coefficient in svm of L2}

\item{tau}{Default param of smooth hinge loss.}

\item{max.steps}{max iteration step}
}
\value{
sparse coefficient
}
\description{
Choice the best feature subset in svm model.
}
\examples{
\dontrun{
data <- gendata(1000,100,5)
bess_svm(data$x, data$y, 5)
}
}
