% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Generate a matrix with n*p have two class.}
\usage{
gendata(n, p, K, sigma = 0.01, seed = 1234)
}
\arguments{
\item{n}{The number of sample.}

\item{p}{The number of dimension.}

\item{K}{The true dimension.}

\item{sigma}{Variance of noise.}

\item{seed}{Random seed.}
}
\value{
A matriX
}
\description{
This function is to generate a matrix with n*p have two class for classification.
}
\examples{
\dontrun{
data <- gendata(1000,100,5)
}
}
