% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{An Rcpp function for Exercise 9.8 (page 278, Statistical Computing with R).}
\usage{
gibbsC(a, b, n)
}
\arguments{
\item{a}{A parameter the bivariate density.}

\item{b}{A parameter the bivariate density.}

\item{n}{A parameter the bivariate density.}
}
\value{
A matrix with n*2.
}
\description{
An Rcpp function for Exercise 9.8 (page 278, Statistical Computing with R).
}
