% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bess_svm.R
\name{bess_svm}
\alias{bess_svm}
\title{Choice the best selection feature.}
\usage{
bess_svm(X, y, T0, alpha = 0.01, tau = 0.05, max.steps = 100)
}
\arguments{
\item{X}{Data Matrix}

\item{y}{True label}

\item{T0}{Nonzero feature number}

\item{alpha}{Default param}

\item{tau}{Default param}

\item{max.steps}{max iteration step}
}
\value{
sparse coefficient
}
\description{
Choice the best selection feature in svm model.
}
\examples{
\dontrun{
data <- gendata(1000,100,5)
bess_svm(data$x, data$y, 5)
}
}
